Fix all-owners.sh when it adds new maintainers
Current script when it finds a new maintainer in one of the repos, adds an empty line since the new maintainer is missing in CNCF's project-maintainers.csv as this is a new maintainer. Instead the fix now prints the github handle and UKNOWN AFFILIATION next to it in the output of the script so that the list can be fixed manually by adding the affiliation and email of the new maintainer.

Example:
Output before the fix:

- Lennart Jern,Ericsson Software Technology,lentzi90,lennart.jern@est.tech
-
- Adam Rozman,Ericsson Software Technology,Rozzii,adam.rozman@est.tech

Output after the fix:
- Lennart Jern,Ericsson Software Technology,lentzi90,lennart.jern@est.tech
- peppi-lotta,----UNKNOWN_AFFILIATION-----
- Adam Rozman,Ericsson Software Technology,Rozzii,adam.rozman@est.tech

Signed-off-by: Kashif Khan <kashif.khan@est.tech>